Move new test to samples_test.go
